refactor: #27 시나리오 아이템 뷰 컴포넌트화
